+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 ### Internal
 
 * Moved JNI build to CMake.
+* Updated Realm Core to 2.0.0-rc4.
 
 ## 1.2.0